+++ examples/generators/content_rewriter/README.md
@@ -15,7 +15,7 @@ The code has been tested on:
 Run the following commands:
 
 ```bash
-cd text_content_manipulation
+cd model
 pip3 install -r requirements.txt
 ```
 
@@ -26,10 +26,12 @@ git clone https://github.com/asyml/texar.git
 cd texar && pip3 install .
 ```
 
-
 ### Downloading the models and data
 
-Before we run the rewriting demo, we need to download models and data from the [link](https://drive.google.com/drive/folders/1jNaJ_R_f89G8xbAC8iwe49Yx_Z-LXr0i) and put the three directories(i.e., data2text, e2e_data, e2e_model) under the `text_content_manipulation/` directory.  
+Before we run the rewriting demo, we need to download models and data from the 
+[link](https://drive.google.com/drive/folders/1jNaJ_R_f89G8xbAC8iwe49Yx_Z-LXr0i) 
+and put the three directories(i.e., data2text, e2e_data, e2e_model) under the 
+`text_content_manipulation/` directory.  
 
 ### Running the example
