+++ examples/generators/content_rewriter/model/README.md
