+++ examples/generators/content_rewriter/model/__init__.py
