+++ examples/generators/content_rewriter/model/config_data_nba.py
