+++ examples/generators/content_rewriter/model/config_model_clean.py
@@ -78,7 +78,6 @@ coverage_rnn_cell = {
     },
 }
 
-
 align_rnn_cell = {
     'type': 'LSTMBlockCell',
     'kwargs': {
