+++ examples/generators/content_rewriter/model/copy_net.py
