+++ examples/generators/content_rewriter/model/requirements.txt
