+++ examples/generators/content_rewriter/model/utils.py
