+++ examples/generators/content_rewriter/rewriter.py
@@ -19,34 +19,15 @@ from forte.common.configuration import Config
 from forte.data.data_pack import DataPack
 from forte.processors.base import PackProcessor
 from ft.onto.base_ontology import Utterance, UtteranceContext
-import os
-import sys
-# class Model:
-#     """This is a dummy model that always return the same string."""
-#
-#     def __init__(self):
-#         self.model_str = "this content is from the model."
-#
-#     def response(self):
-#         # Pretend the model is running for 5 seconds.
-#         print('model is computing...')
-#         import time
-#         time.sleep(5)
-#         return self.model_str
+from examples.generators.content_rewriter. \
+    model.manip import Rewriter
 
 
 class ContentRewriter(PackProcessor):
     def initialize(self, resources: Resources, configs: Config):
         # pylint: disable=attribute-defined-outside-init
 
-        # Make sure the initialize model here.
-        # Note: here we first need to set the path for reading data later.
-        # print('[info]{}'.format(os.getcwd()))
-
-        os.chdir('../../forte-rewriter/examples/generators/content_rewriter')
-
-        from examples.generators.content_rewriter.text_content_manipulation.manip import Rewriter
-        #initialize model
+        # initialize model
         self.model = Rewriter()
         self.model.load_model()
 
@@ -77,5 +58,4 @@ class ContentRewriter(PackProcessor):
         print('The utterance is:')
         print(utterance.text)
 
-        # self.new_utternace(input_pack, self.model.eval_epoch(self.model.sess, self.model.summary_writer, 'test'), 'ai')
         self.new_utternace(input_pack, self.model.eval_epoch('test'), 'ai')
